CTFE engine: rename copy → copy_intrinsic, move to intrinsics.rs
